create SearchQuery and use it in LocalQuery
