+++ react/SearchQuery.js
@@ -1,0 +1,2 @@
+import SearchQuery from './components/SearchQuery'
+export default SearchQuery
