+++ react/components/SearchQuery.js
@@ -1,0 +1,104 @@
+import { zip, split, head, join, tail } from 'ramda'
+import React, { useMemo } from 'react'
+import { Query } from 'react-apollo'
+import { productSearchV2 } from 'vtex.store-resources/Queries'
+
+const DEFAULT_PAGE = 1
+
+const QUERY_SEPARATOR = '/'
+const MAP_SEPARATOR = ','
+
+const splitQuery = split(QUERY_SEPARATOR)
+const splitMap = split(MAP_SEPARATOR)
+const joinQuery = join(QUERY_SEPARATOR)
+const joinMap = join(MAP_SEPARATOR)
+
+const includeFacets = (map, query) =>
+  !!(map && map.length > 0 && query && query.length > 0)
+
+const useFacetsArgs = (query, map) => {
+  return useMemo(() => {
+    const queryArray = splitQuery(query)
+    const mapArray = splitMap(map)
+    const queryAndMap = zip(queryArray, mapArray)
+    const relevantArgs = [
+      head(queryAndMap),
+      ...tail(queryAndMap).filter(
+        ([_, tupleMap]) => tupleMap === 'c' || tupleMap === 'ft'
+      ),
+    ]
+    const { finalMap, finalQuery } = relevantArgs.reduce(
+      (accumulator, [tupleQuery, tupleMap]) => {
+        accumulator.finalQuery.push(tupleQuery)
+        accumulator.finalMap.push(tupleMap)
+        return accumulator
+      },
+      { finalQuery: [], finalMap: [] }
+    )
+    const facetQuery = joinQuery(finalQuery)
+    const facetMap = joinMap(finalMap)
+    return {
+      facetQuery,
+      facetMap,
+      withFacets: includeFacets(facetMap, facetQuery),
+    }
+  }, [map, query])
+}
+
+const SearchQuery = ({
+  maxItemsPerPage,
+  query,
+  map,
+  orderBy,
+  priceRange,
+  hideUnavailableItems,
+  pageQuery,
+  children,
+}) => {
+  const page = pageQuery ? parseInt(pageQuery) : DEFAULT_PAGE
+  const from = (page - 1) * maxItemsPerPage
+  const to = from + maxItemsPerPage - 1
+  const facetsArgs = useFacetsArgs(query, map)
+  const variables = useMemo(() => {
+    return {
+      query,
+      map,
+      orderBy,
+      priceRange,
+      from,
+      to,
+      hideUnavailableItems,
+      ...facetsArgs,
+    }
+  }, [
+    query,
+    map,
+    orderBy,
+    priceRange,
+    from,
+    to,
+    hideUnavailableItems,
+    facetsArgs,
+  ])
+  console.log('teste === variables: ', variables)
+  const extraParams = useMemo(() => {
+    return {
+      ...variables,
+      page,
+    }
+  }, [variables, page])
+  return (
+    <Query
+      query={productSearchV2}
+      variables={variables}
+      notifyOnNetworkStatusChange
+      partialRefetch
+    >
+      {searchQuery => {
+        return children(searchQuery, extraParams)
+      }}
+    </Query>
+  )
+}
+
+export default SearchQuery
